+++ migrations/sqlite/20260913000000_rename_redirects_stele_name_to_fonds_name.sql
@@ -1,3 +1,0 @@
-ALTER TABLE redirects RENAME COLUMN stele_name TO fonds_name;
-
-PRAGMA optimize;