+++ tests/basic/mod.rs
@@ -1,2 +1,3 @@
 mod archive_test;
 mod gitrepo_test;
+mod publication_test;